Add choice to use DB or memory-based variable

+++ src/main/kotlin/com/kangdroid/server/remover/RemoverService.kt
@@ -70,6 +70,10 @@ class RemoverService {
 
     // DB --> Local[check invalid entry]
     fun initDB(dbList: List<TrashDataResponseDto>) {
+        if (settings.lowMemoryOption == false) {
+            return
+        }
+
         for (target in dbList) {
             if (!File(target.trashFileDirectory).exists()) {
                 dataService.deleteById(target.id)
